crud competition scramble done

+++ app/Http/Controllers/Admin/CompetitionItemController.php
@@ -5,6 +5,7 @@ namespace App\Http\Controllers\Admin;
 use App\Http\Controllers\Controller;
 use App\Http\Requests\CompetitionItem\StoreCompetitionItem;
 use App\Http\Requests\CompetitionItem\UpdateCompetitionItem;
+use App\Models\Competition;
 use App\Models\CompetitionItem;
 use App\Models\CompetitionRound;
 use App\Models\CubeCategory;
@@ -16,16 +17,25 @@ class CompetitionItemController extends Controller
     public function store(StoreCompetitionItem $request) 
     {
         $item = new CompetitionItem($request->validated());
+        $competition = Competition::findOrFail($request->competition_id);
+        $competition_date = Competition::where('id', $request->competition_id)
+                                    ->whereDate('date_start', '<=', $request->date)
+                                    ->whereDate('date_end', '>=', $request->date)
+                                    ->first();
 
         try {
-            if($request->hasFile('scramble_img')) {
-                $uploadedImage = $request->file('scramble_img')->store('uploads/scrambles', 'public');
-                $item->scramble_img = $uploadedImage;
-            }  
-
-            $item->save();
+            if($competition_date) {
+                if($request->hasFile('scramble_img')) {
+                    $uploadedImage = $request->file('scramble_img')->store('uploads/scrambles', 'public');
+                    $item->scramble_img = $uploadedImage;
+                }  
+    
+                $item->save();
+                session()->flash('success', 'Competition scramble created successfully');;
+            } else {
+                session()->flash('error', 'Date must be between ' . date_format(date_create($competition->date_start), "d M Y") . ' and ' . date_format(date_create($competition->date_end), "d M Y"));
+            }
             
-            session()->flash('success', 'Competition scramble created successfully');;
         } catch (\Exception $e) {
             session()->flash('error', 'Competition scramble created failed');
         }
@@ -41,28 +51,55 @@ class CompetitionItemController extends Controller
         return response()->json([$item, $round_name, $category_name]);
     }
 
-    public function update(UpdateCompetitionItem $request)
+    public function update(UpdateCompetitionItem $request, $id)
     {
-        $item = CompetitionItem::findOrFail($request->competition_item_id);
-        
-        try {
-            $fileName = $item->scramble_img;
-            if($request->hasFile('scramble_img')) {
-                if(File::exists('storage/' . $fileName)) {
-                    File::delete('storage/' . $fileName);
-                }
-                $fileName = $request->file('scramble_img')->store('uploads/scrambles', 'public');
-            }
+        $item = CompetitionItem::findOrFail($id);
+        $competition = Competition::findOrFail($request->competition_id);
+        $competition_date = Competition::where('id', $request->competition_id)
+                                    ->whereDate('date_start', '<=', $request->date)
+                                    ->whereDate('date_end', '>=', $request->date)
+                                    ->first();
 
-            $item->competition_round_id = $request->competition_round_id;
-            $item->cube_category_id = $request->cube_category_id;
-            $item->date = $request->date;
-            $item->scramble_img = $fileName;
-            $item->update();
-            
-            session()->flash('success', 'Competition scramble updated successfully');
+        try {
+            if($competition_date) {
+                $fileName = $item->scramble_img;
+                if($request->hasFile('scramble_img')) {
+                    if(File::exists('storage/' . $fileName)) {
+                        File::delete('storage/' . $fileName);
+                    }
+                    $fileName = $request->file('scramble_img')->store('uploads/scrambles', 'public');
+                }
+
+                $item->competition_round_id = $request->competition_round_id;
+                $item->cube_category_id = $request->cube_category_id;
+                $item->date = $request->date;
+                $item->scramble_img = $fileName;
+                $item->update();
+                
+                session()->flash('success', 'Competition scramble updated successfully');
+            } else {
+                session()->flash('error', 'Date must be between ' . date_format(date_create($competition->date_start), "d M Y") . ' and ' . date_format(date_create($competition->date_end), "d M Y"));
+            }
         } catch (\Exception $e) {
             session()->flash('error', 'Competition scramble updated failed');
+        }
+    }
+
+    public function destroy($id)
+    {
+        $item = CompetitionItem::findOrFail($id);
+
+        try {    
+            $path = $item->scramble_img;        
+            if(File::exists('storage/' . $path)) {
+                File::delete('storage/' . $path);
+            }
+
+            $item->delete();
+            
+            session()->flash('success', 'Competition scramble successfully');
+        } catch (\Exception $e) {
+            return back()->with('error', 'Competition scramble failed');
         }
     }
 }
